+++ CHANGELOG.md
@@ -9,6 +9,9 @@
 - On Windows, prevent ghost window from showing up in the taskbar after either several hours of use or restarting `explorer.exe`.
 - On macOS, fix issue where `ReceivedCharacter` was not being emitted during some key repeat events.
 - On Wayland, load cursor icons `hand2` and `hand1` for `CursorIcon::Hand`.
+- **Breaking:** On Wayland, Theme trait and its support types are dropped.
+- On Wayland, bump `smithay-client-toolkit` to 0.15.
+
 
 # 0.25.0 (2021-05-15)